Extract common behaviour to BaseApiController
